+++ src/Aws/S3.php
@@ -46,7 +46,8 @@ class S3 extends Client
 
     public function createWebsite($indexDoc = 'index.html', $errorDoc = 'error.html') {
         try {
-            $result = $this->client->putBucketWebsite([
+            // enable static web hosting
+            $this->client->putBucketWebsite([
                 'Bucket' => $this->bucket, // REQUIRED
                 'WebsiteConfiguration' => [ // REQUIRED
                     'ErrorDocument' => [
@@ -57,6 +58,13 @@ class S3 extends Client
                     ],
                 ],
             ]);
+
+            // remove public access block
+            $this->deletePublicAccessBlock();
+
+            // add policy to enable public access (GET)
+            $this->enablePublicAccess();
+
         } catch(\AwsException $e) {
             throw new \Exception("Unable to create website");
         }